requirements.txt
requests
pandas
flask
scikit-learn
joblib
pytest
python-dotenv
